+++ projeto-obra/src/main/java/com/projeto/obra/entities/ObraLocalizacao.java
@@ -1,7 +1,5 @@
 package com.projeto.obra.entities;
 
-import java.math.BigInteger;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -13,7 +11,7 @@ public class ObraLocalizacao {
     
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private BigInteger id;
+    private Integer id;
     @ManyToOne
     private Obra obraId;
     private String estado;
@@ -24,10 +22,10 @@ public class ObraLocalizacao {
     public ObraLocalizacao() {
     }
     
-    public BigInteger getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(BigInteger id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public Obra getObraId() {
